Refactor all legacy config to new ones
