--- src/main/java/cn/dreeam/surf/listener/ListenerStackedTotem.java
@@ -1,27 +1,0 @@
-package cn.dreeam.surf.listener;
-
-import cn.dreeam.surf.config.Config;
-import cn.dreeam.surf.util.item.ItemUtil;
-import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.entity.EntityDamageEvent;
-
-public class ListenerStackedTotem implements Listener {
-
-    @EventHandler
-    private void onDamage(EntityDamageEvent event) {
-        if (!Config.stackedTotemRevertAsOneEnabled) return;
-
-        // Check when player gets damage or make damage to others
-        if (event.getEntity() instanceof Player entity) {
-            // TODO, move to as an option
-            entity.getInventory().forEach(i -> {
-                        if (i != null && ItemUtil.isIllegalTotem(i)) {
-                            i.setAmount(1);
-                        }
-                    }
-            );
-        }
-    }
-}
